Updated Data29012021

+++ AdminPanel/AdminPanel/bin/Release/Publish/CustomScripts/jsAreaMst.js
@@ -7,18 +7,28 @@
     area_list();
 
     $("#btnSave").click(function () {
-        var requestData = '[{"country_id":"' + $("#cmboCountryList").val() + '","zone_id":"' + $("#cmboZoneList").val() + '","state_id":"' + $("#cmboStateList").val() + '","city_id":"' + $("#cmboCityList").val() + '","area_name":"' + $("#txtAreaName").val() + '","user_id":"11"}]';
-        $.ajax({
-            type: "POST",
-            url: "https://trial.spyderxindia.com/api/areamst",
-            data: JSON.stringify(requestData),
-            contentType: "application/json",
-            //datatype: "json",
-            success: function (responseFromServer) {
-                alert(responseFromServer);
-                area_list();
-            }
-        });
+
+        var re = /^[a-zA-Z]+$/;
+        if (!re.test($("#txtAreaName").val())) {
+
+            alert("Enter valid Area name...!");
+
+        }
+        else {
+
+            var requestData = '[{"country_id":"' + $("#cmboCountryList").val() + '","zone_id":"' + $("#cmboZoneList").val() + '","state_id":"' + $("#cmboStateList").val() + '","city_id":"' + $("#cmboCityList").val() + '","area_name":"' + $("#txtAreaName").val() + '","user_id":"11"}]';
+            $.ajax({
+                type: "POST",
+                url: "https://trial.spyderxindia.com/api/areamst",
+                data: JSON.stringify(requestData),
+                contentType: "application/json",
+                //datatype: "json",
+                success: function (responseFromServer) {
+                    alert(responseFromServer);
+                    area_list();
+                }
+            });
+        }
     });
 
 
@@ -39,9 +49,6 @@
             }
         });
     });
-
-
-
 });
 
 
@@ -199,9 +206,9 @@ function update(id) {
         contentType: "application/json",
         //datatype: "json",
         success: function (responseFromServer) {
-            alert(responseFromServer);
+            //alert(responseFromServer);
             var jsonData = eval('(' + responseFromServer + ')');
-            alert(jsonData[0].index_id);
+           // alert(jsonData[0].index_id);
             var tbl = "<table>";
             for (var iCount = 0; iCount < jsonData.length; iCount++) {
                 $("#txtAreaId").val(jsonData[0].index_id);
@@ -219,20 +226,22 @@ function update(id) {
 
 
 function del(id) {
-    $.ajax({
-        //url: "https://trial.spyderxindia.com/api/citymst/3",
-        type: "DELETE",
-        //crossOrigin: true,
-        url: "https://trial.spyderxindia.com/api/areamst/" + id,
-        //data: JSON.stringify(num),
-        contentType: "application/json",
-        //datatype: "json",
-        success: function (responseFromServer) {
-            alert(responseFromServer);
-            area_list();
+    if (confirm('Are you sure ?')) {
+        $.ajax({
+            //url: "https://trial.spyderxindia.com/api/citymst/3",
+            type: "DELETE",
+            //crossOrigin: true,
+            url: "https://trial.spyderxindia.com/api/areamst/" + id,
+            //data: JSON.stringify(num),
+            contentType: "application/json",
+            //datatype: "json",
+            success: function (responseFromServer) {
+                alert(responseFromServer);
+                area_list();
 
 
-        }
-    });
+            }
+        });
+    }
 
 }
