+++ AdminPanel/AdminPanel/bin/Release/Publish/CustomScripts/jsSubscriptionMst.js
@@ -1,15 +1,15 @@
 ﻿$(document).ready(function () {
     //alert("Hello");
     country_list();
-    zone_list();
-    state_list();
+    subscription_plan_list();
+    customer_list();
     subscription_list();
 
     $("#btnSave").click(function () {
-        var requestData = '[{"country_id":"' + $("#cmboCountryList").val() + '","zone_id":"' + $("#cmboZoneList").val() + '","state_id":"' + $("#cmboStateList").val() + '","city_name":"' + $("#txtCityName").val() + '","user_id":"11"}]';
+        var requestData = '[{"country_id":"' + $("#cmboCountryList").val() + '","plan_id":"' + $("#cmboSubscriptionPlanList").val() + '","subscription_date":"' + $("#txtSubscription_Date").val() + '","subscription_expire_date":"' + $("#txtSubscription_Expire_Date").val() + '","customer_id":"' + $("#cmboCustomerList").val() + '","user_id":"11"}]';
         $.ajax({
             type: "POST",
-            url: "https://trial.spyderxindia.com/api/citymst",
+            url: "https://trial.spyderxindia.com/api/subscriptionmst",
             data: JSON.stringify(requestData),
             contentType: "application/json",
             //datatype: "json",
@@ -24,10 +24,10 @@
 
     $("#btnUpdate").click(function () {
 
-        var requestData = '[{"country_id":"' + $("#cmboCountryList").val() + '","zone_id":"' + $("#cmboZoneList").val() + '","state_id":"' + $("#cmboStateList").val() + '","city_name":"' + $("#txtCityName").val() + '","user_id":"11"}]';
+        var requestData = '[{"country_id":"' + $("#cmboCountryList").val() + '","plan_id":"' + $("#cmboSubscriptionPlanList").val() + '","subscription_date":"' + $("#txtSubscription_Date").val() + '","subscription_expire_date":"' + $("#txtSubscription_Expire_Date").val() + '","customer_id":"' + $("#cmboCustomerList").val() + '","user_id":"11"}]';
         $.ajax({
             type: "PUT",
-            url: "https://trial.spyderxindia.com/api/citymst/" + $("#txtCityId").val(),
+            url: "https://trial.spyderxindia.com/api/subscriptionmst/" + $("#txtSubscription_Id").val(),
             data: JSON.stringify(requestData),
             contentType: "application/json",
             //datatype: "json",
@@ -47,13 +47,13 @@
 
 
 
-function city_list() {
+function subscription_list() {
 
     $.ajax({
-        //url: "https://trial.spyderxindia.com/api/statemst/3",
+        //url: "https://trial.spyderxindia.com/api/subscriptionmst/3",
         type: "GET",
         //crossOrigin: true,
-        url: "https://trial.spyderxindia.com/api/citymst",
+        url: "https://trial.spyderxindia.com/api/subscriptionmst",
         //data: JSON.stringify(num),
         contentType: "application/json",
         //datatype: "json",
@@ -62,36 +62,48 @@ function city_list() {
             var jsonData = eval('(' + responseFromServer + ')');
             //alert(jsonData[0].index_id);
             var tbl = "<table width='100%'>";
-            tbl += "<tr><th>Id</th><th>City Name</th><th>Delete</th><th>Update</th></tr>";
+            tbl += "<tr><th>Id</th><th>Country Name</th><th>Plan Name</th><th>Sub Date</th><th>Sub Expire Date</th><th>Customer Name</th><th>Delete</th><th>Update</th></tr>";
             for (var iCount = 0; iCount < jsonData.length; iCount++) {
                 tbl += "<tr>";
-                tbl += "<td width='10%'>";
+                tbl += "<td>";
                 tbl += iCount + 1;
                 tbl += "</td>";
-                tbl += "<td width='40%'>";
-                tbl += jsonData[iCount].city_name;
+                tbl += "<td>";
+                tbl += jsonData[iCount].country_name;
                 tbl += "</td>";
-                tbl += "<td width='25%'>";
+                tbl += "<td>";
+                tbl += jsonData[iCount].plan_id;
+                tbl += "</td>";
+                tbl += "<td>";
+                tbl += jsonData[iCount].subscription_date;
+                tbl += "</td>";
+                tbl += "<td>";
+                tbl += jsonData[iCount].subscription_expire_date;
+                tbl += "</td>";
+                tbl += "<td>";
+                tbl += jsonData[iCount].customer_id;
+                tbl += "</td>";
+                tbl += "<td>";
                 tbl += "<i type='button' class='fa fa-trash btn-danger'  onclick='del(" + jsonData[iCount].index_id + ")'></i>";
                 tbl += "</td>";
-                tbl += "<td width='25%'>";
+                tbl += "<td>";
                 tbl += "<i type='button' class='fa fa-edit btn-warning' onclick='update(" + jsonData[iCount].index_id + ")'></i>";
                 tbl += "</td>";
                 tbl += "</tr>";
             }
             tbl += "</table>";
-            $("#lstCityList").html(tbl);
+            $("#lstSubscriptionList").html(tbl);
 
         }
     });
 
 }
 
-function state_list() {
+function customer_list() {
     $.ajax({
         type: "GET",
         //crossOrigin: true,
-        url: "https://trial.spyderxindia.com/api/statemst",
+        url: "https://trial.spyderxindia.com/api/customermst",
         //data: JSON.stringify(num),
         contentType: "application/json",
         //datatype: "json",
@@ -102,21 +114,21 @@ function state_list() {
             var cmbOptions = "";
             for (var iCount = 0; iCount < jsonData.length; iCount++) {
                 cmbOptions += "<option value=" + jsonData[iCount].index_id + ">";
-                cmbOptions += jsonData[iCount].state_name;
+                cmbOptions += jsonData[iCount].customer_name;
                 cmbOptions += "</option>";
             }
-            $("#cmboStateList").html(cmbOptions);
+            $("#cmboCustomerList").html(cmbOptions);
 
         }
     });
 
 }
 
-function zone_list() {
+function subscription_plan_list() {
     $.ajax({
         type: "GET",
         //crossOrigin: true,
-        url: "https://trial.spyderxindia.com/api/zonemst",
+        url: "https://trial.spyderxindia.com/api/subscriptionplanmst",
         //data: JSON.stringify(num),
         contentType: "application/json",
         //datatype: "json",
@@ -127,10 +139,10 @@ function zone_list() {
             var cmbOptions = "";
             for (var iCount = 0; iCount < jsonData.length; iCount++) {
                 cmbOptions += "<option value=" + jsonData[iCount].index_id + ">";
-                cmbOptions += jsonData[iCount].zone_name;
+                cmbOptions += jsonData[iCount].subscription_plan_name;
                 cmbOptions += "</option>";
             }
-            $("#cmboZoneList").html(cmbOptions);
+            $("#cmboSubscriptionPlanList").html(cmbOptions);
 
         }
     });
@@ -165,21 +177,21 @@ function country_list() {
 
 function update(id) {
     $.ajax({
-        //url: "https://trial.spyderxindia.com/api/statemst/3",
+        //url: "https://trial.spyderxindia.com/api/subscriptionmst/3",
         type: "GET",
         //crossOrigin: true,
-        url: "https://trial.spyderxindia.com/api/citymst/" + id,
+        url: "https://trial.spyderxindia.com/api/subscriptionmst/" + id,
         //data: JSON.stringify(num),
         contentType: "application/json",
         //datatype: "json",
         success: function (responseFromServer) {
-            alert(responseFromServer);
+           // alert(responseFromServer);
             var jsonData = eval('(' + responseFromServer + ')');
-            alert(jsonData[0].index_id);
+           // alert(jsonData[0].index_id);
             var tbl = "<table>";
             for (var iCount = 0; iCount < jsonData.length; iCount++) {
-                $("#txtCityId").val(jsonData[0].index_id);
-                $("#txtCityName").val(jsonData[0].city_name);
+                $("#txtSubscriptionId").val(jsonData[0].index_id);
+                $("#txtCountryName").val(jsonData[0].country_name);
 
 
             }
@@ -193,20 +205,21 @@ function update(id) {
 
 
 function del(id) {
-    $.ajax({
-        //url: "https://trial.spyderxindia.com/api/statemst/3",
-        type: "DELETE",
-        //crossOrigin: true,
-        url: "https://trial.spyderxindia.com/api/citymst/" + id,
-        //data: JSON.stringify(num),
-        contentType: "application/json",
-        //datatype: "json",
-        success: function (responseFromServer) {
-            alert(responseFromServer);
-           subscription_list();
+    if (confirm('Are you sure ?')) {
+        $.ajax({
+            //url: "https://trial.spyderxindia.com/api/subscriptionmst/3",
+            type: "DELETE",
+            //crossOrigin: true,
+            url: "https://trial.spyderxindia.com/api/subscriptionmst/" + id,
+            //data: JSON.stringify(num),
+            contentType: "application/json",
+            //datatype: "json",
+            success: function (responseFromServer) {
+                alert(responseFromServer);
+                subscription_list();
 
 
-        }
-    });
-
+            }
+        });
+    }
 }
